requirements.txt
openai==1.34.0
python-dotenv==1.0.1
pyaudio
websockets